+++ paddle/phi/kernels/gpu/graph_send_ue_recv_funcs.h
@@ -185,7 +185,8 @@ __global__ void ManipulateMeanGradCUDAKernelForMulX(const T* out_grad_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     const T* out_grad_off = out_grad_data + src * l_len;
     const T* e_off = e_data + ty * r_len;
@@ -219,7 +220,8 @@ __global__ void ManipulateSumGradCUDAKernelForAddE(const T* out_grad_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     T* e_grad_off = e_grad + ty * r_len;
     const T* out_grad_off = out_grad_data + dst * out_len;
@@ -255,7 +257,8 @@ __global__ void ManipulateSumGradCUDAKernelForMulE(const T* x_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     const T* x_off = x_data + src * l_len;
     T* e_grad_off = e_grad + ty * r_len;
@@ -289,7 +292,8 @@ __global__ void ManipulateMeanGradCUDAKernelForAddE(const T* out_grad_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     T* e_grad_off = e_grad + ty * r_len;
     const T* out_grad_off = out_grad_data + dst * out_len;
@@ -327,7 +331,8 @@ __global__ void ManipulateMeanGradCUDAKernelForMulE(const T* x_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     const T* x_off = x_data + src * l_len;
     T* e_grad_off = e_grad + ty * r_len;
@@ -370,7 +375,8 @@ __global__ void ManipulateMinMaxGradCUDAKernelForAdd(const T* x_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     const T* x_off = x_data + dst * x_len;
     const T* e_off = e_data + ty * e_len;
@@ -418,7 +424,8 @@ __global__ void ManipulateMinMaxGradCUDAKernelForMul(const T* x_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     const T* x_off = x_data + dst * x_len;
     const T* e_off = e_data + ty * e_len;
