+++ paddle/phi/kernels/gpu/send_uv_kernel.cu
@@ -49,7 +49,8 @@ __global__ void GraphSendUVCUDAKernel(const T* x_data,
     int64_t tx =
         static_cast<int64_t>(blockIdx.x) * static_cast<int64_t>(blockDim.x) +
         static_cast<int64_t>(threadIdx.x);
-    int64_t stride_x = blockDim.x * gridDim.x;
+    int64_t stride_x =
+        static_cast<int64_t>(blockDim.x) * static_cast<int64_t>(gridDim.x);
 
     const T* x_off = x_data + src * x_len;
     const T* y_off = y_data + dst * y_len;
